+++ React/app/src/components/fullPageListing.tsx
@@ -3,41 +3,11 @@ import React, { useEffect, useState } from 'react';
 
 import './FullPageListing.scss'
 import PrescriptionGrid from './primitives/PrescriptionGrid';
+import {DetailedPosting } from '../interfaces';
 
 // props interface
 interface props {
     listingNumber: number
-}
-
-// return data interfaces
-interface DetailedPosting {
-    comment: string,
-    contact?: string | null,
-    flagged: boolean,
-    location: string,
-    pictures? : Image[],
-    postNumb: number,
-    prescription?: Prescription | null,
-    pseudoPrescription?: number | null,
-    user: string,
-}
-
-interface EyePrescription {
-    sphere: number | null;
-    cylinder: number | null;
-    axis: number | null;
-    prism: number | null;
-    base: string | null;
-}
-  
-interface Prescription {
-    leftEye: EyePrescription;
-    rightEye: EyePrescription;
-}
-
-interface Image {
-    id: number,
-    imagePath: string,
 }
 
 // class
@@ -67,7 +37,7 @@ const FullPageListing: React.FC<props> = ({listingNumber}) => {
             {error? (
                 <div>Error fetching post data: {error}</div>
             ):(
-                <PrescriptionGrid pseudoPrescription={details?.pseudoPrescription ?? null} prescription={details?.prescription ?? null}></PrescriptionGrid>    
+                <PrescriptionGrid prescription={details?.prescription ?? null}></PrescriptionGrid>    
             )}
             {/* {error? (
                 <div>Error fetching post data: {error}</div>
