+++ React/app/src/interfaces.tsx
@@ -1,4 +1,4 @@
-// prescription interface
+// interface for prescription
 interface EyePrescription {
     sphere: number | null;
     cylinder: number | null;
@@ -10,4 +10,38 @@ interface EyePrescription {
 export interface Prescription {
     leftEye: EyePrescription;
     rightEye: EyePrescription;
+}
+
+// interfaces for different sets of post information
+interface Image {
+    id: number,
+    imagePath: string,
+}
+
+export interface DetailedPosting {
+    comment: string,
+    contact?: string | null,
+    flagged: boolean,
+    location: string,
+    pictures? : Image[],
+    postNumb: number,
+    prescription?: Prescription | null,
+    pseudoPrescription?: number | null,
+    user: string,
+}
+
+export interface Posting {
+    postNumb: number,
+    location: string,
+    sphere: number,
+    flagged: boolean,
+    imageCard: string,
+}
+
+export interface NewPostForm {
+    prescription: Prescription;
+    pseudoPrescription: number | null;
+    comment: string | null;
+    location: string;
+    contact: string;
 }
